86cz0ug47 - fetched required attributes of property listing
